replace old LanguageLookup with NewLanguageLookup
