+++ SIL.WritingSystems.Tests/SIL.WritingSystems.Tests.csproj
@@ -205,7 +205,6 @@
     </Compile>
     <Compile Include="AddSortKeysToXmlTests.cs" />
     <Compile Include="CharacterSetDefinitionCloneableTests.cs" />
-    <Compile Include="NewLanguageLookupTests.cs" />
     <Compile Include="CollationDefinitionCloneableTests.cs" />
     <Compile Include="IcuRulesCollationDefinitionTests.cs" />
     <Compile Include="IetfLanguageTagTests.cs" />
